Deleted a MOUNTAIN of dead code and refactored some. Still more to come. Start new run at picking, but ran out of time, so will have to be finished later.

+++ src/app/MasterNode.h
@@ -102,7 +102,7 @@ namespace viscom {
         bool UpdateTuioCursor(TUIO::TuioCursor *tcur) override;
         bool RemoveTuioCursor(TUIO::TuioCursor *tcur) override;
 
-        glm::vec2 GetCirclePos(glm::vec2 center, float radius, float angle);
+        glm::vec2 GetCirclePos(glm::vec2 center, float radius, int angle);
         int viewAngle = 90;
     };
 }
